Convert Defensives into an implementation of an abstract Utilities analyser
