+++ src/parser/core/modules/DISPLAY_ORDER.ts
@@ -9,6 +9,7 @@ export enum DISPLAY_ORDER {
 	POSITIONALS = 120,
 	OVERHEAL = 124,
 	DEFENSIVES = 125,
+	UTILITIES = 126,
 	TIMELINE = 130,
 	EVENTS_VIEW = 140,
 	CHANGELOG = 150,
